pyproject.toml
[tool.poetry]
name = "linkedin-search-agent"
version = "0.1.0"
description = "LLM based agent for autonomously searching specific linkedIn profiles as promppted."
authors = ["Ed Mwanza"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openai = "^1.35.7"
torch = "^2.3.1"
langchain = "^0.2.6"
setuptools = "^70.1.1"
langchain-community = "^0.2.6"
pydantic = "^2.7.4"
google-search-results = "^2.4.2"
black = "^24.4.2"
langchain-openai = "^0.1.11"
autopep8 = "^2.3.1"
langchainhub = "^0.1.20"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
